// data/github/current/ui/packages/pull-request-viewer/components/header/__tests__/__generated__/PullRequestHeaderTestQuery.graphql.ts
export type PullRequestHeaderTestQuery$variables = {
  endOid?: string | null | undefined;
  number: number;
  owner: string;
  repo: string;
  singleCommitOid?: string | null | undefined;
  startOid?: string | null | undefined;
};
export type PullRequestHeaderTestQuery$data = {
  readonly repository: {
    readonly " $fragmentSpreads": FragmentRefs<"PullRequestHeaderWrapper_repository">;
  } | null | undefined;
  readonly viewer: {
    readonly " $fragmentSpreads": FragmentRefs<"PullRequestHeaderWrapper_user">;
  };
};
export type PullRequestHeaderTestQuery = {
  response: PullRequestHeaderTestQuery$data;
  variables: PullRequestHeaderTestQuery$variables;
};